+++ R/conv-methods.R
@@ -1,4 +1,10 @@
 #' @describeIn conv Extracts the convergence status of the model with respect to
+#'   the parameter.
+setMethod("conv", "CEL", function(object, ...) {
+  getOptim(object)$convergence
+})
+
+#' @describeIn conv Extracts the convergence status of the model with respect to
 #'   the Lagrange multiplier.
 #' @srrstats {RE4.7} `conv()` method extracts information on convergence status.
 setMethod("conv", "EL", function(object, ...) {
@@ -6,8 +12,8 @@ setMethod("conv", "EL", function(object, ...) {
 })
 
 #' @describeIn conv Extracts the convergence status of the model with respect to
-#'   the parameter.
-setMethod("conv", "CEL", function(object, ...) {
+#'   the parameter (or the Lagrange multiplier if `lhs` is `NULL`).
+setMethod("conv", "ELT", function(object, ...) {
   getOptim(object)$convergence
 })
 
@@ -15,10 +21,4 @@ setMethod("conv", "CEL", function(object, ...) {
 #'   documentation of \linkS4class{EL} and \linkS4class{CEL}.
 setMethod("conv", "SummaryLM", function(object, ...) {
   object@convergence
-})
-
-#' @describeIn conv Extracts the convergence status of the model with respect to
-#'   the parameter (or the Lagrange multiplier if `lhs` is `NULL`).
-setMethod("conv", "ELT", function(object, ...) {
-  getOptim(object)$convergence
 })