+++ R/el_mean.R
@@ -32,7 +32,7 @@
 #'   “Empirical Likelihood Ratio Confidence Regions.”
 #'   \emph{The Annals of Statistics}, 18(1), 90--120.
 #'   \doi{10.1214/aos/1176347494}.
-#' @seealso \linkS4class{EL}, [el_control()], [el_eval()], [elt()]
+#' @seealso \linkS4class{EL}, [elt()], [el_eval()], [el_control()]
 #' @examples
 #' ## Scalar mean
 #' set.seed(414)